Remove stuff that's not being used yet.
Eventually, I want to add the Re4son kernel to this script.  To do that, I'll need to reboot the system while the script is running, and then I'd like to just automatically pick back up.  Since Re4son hasn't been added to this build yet, there's no "reason" for this script to try to restart after a reboot.